--- src/SimpleLED/SimpleLED.sln
@@ -1,22 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 2013
-VisualStudioVersion = 12.0.21005.1
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SimpleLED", "SimpleLED\SimpleLED.csproj", "{EE2ED13F-87A1-4941-BCBC-E3BC6CF9CE7F}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{EE2ED13F-87A1-4941-BCBC-E3BC6CF9CE7F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{EE2ED13F-87A1-4941-BCBC-E3BC6CF9CE7F}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{EE2ED13F-87A1-4941-BCBC-E3BC6CF9CE7F}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{EE2ED13F-87A1-4941-BCBC-E3BC6CF9CE7F}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
