--- src/SimpleLED/SimpleLED/MainWindow.xaml
@@ -1,51 +1,0 @@
-﻿<Window x:Class="SimpleLED.MainWindow"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        Title="SimpleLED" Height="350" Width="550" ResizeMode="NoResize">
-    <Grid>
-        <Grid.ColumnDefinitions>
-            <ColumnDefinition Width="50"/>
-            <ColumnDefinition Width="100"/>
-            <ColumnDefinition/>
-        </Grid.ColumnDefinitions>
-        <Grid.RowDefinitions>
-            <RowDefinition Height="25"/>
-            <RowDefinition Height="75"/>
-            <RowDefinition Height="25"/>
-            <RowDefinition Height="100"/>
-            <RowDefinition Height="65"/>
-            <RowDefinition/>
-        </Grid.RowDefinitions>
-        <Label Content="Port : " Grid.ColumnSpan="3"/>
-        <ComboBox x:Name="comBox" Grid.Column="1" Text="Port">
-            <ComboBoxItem Name="com1" Content="COM1"/>
-            <ComboBoxItem Name="com2" Content="COM2"/>
-            <ComboBoxItem Name="com3" Content="COM3"/>
-            <ComboBoxItem Name="com4" Content="COM4"/>
-            <ComboBoxItem Name="com5" Content="COM5"/>
-            <ComboBoxItem Name="com6" Content="COM6"/>
-            <ComboBoxItem Name="com7" Content="COM7"/>
-            <ComboBoxItem Name="com8" Content="COM8"/>
-            <ComboBoxItem Name="com9" Content="COM9"/>
-            <ComboBoxItem Name="com10" Content="COM10"/>
-        </ComboBox>
-        <Label Content="Red: " Grid.Row="1" VerticalAlignment="Top"/>
-        <Label Content="Green: " Grid.Row="1" VerticalAlignment="Center" Margin="0,25,0,24"/>
-        <Label Content="Blue: " Grid.Row="1" VerticalAlignment="Bottom"/>
-        <Slider x:Name="redSlider" Grid.Column="1" Grid.ColumnSpan="2" Grid.Row="1" VerticalAlignment="Top" Margin="0,5,0,0" ValueChanged="ColorChanged" Maximum="255"/>
-        <Slider x:Name="greenSlider" Grid.Column="1" Grid.ColumnSpan="2" Grid.Row="1" VerticalAlignment="Center" Margin="0,31,0,26" ValueChanged="ColorChanged" Maximum="255"/>
-        <Slider x:Name="blueSlider" Grid.Column="1" Grid.ColumnSpan="2" Grid.Row="1" VerticalAlignment="Bottom" ValueChanged="ColorChanged" Maximum="255"/>
-        <Label Content="Color preview : " Grid.Row="2" Grid.Column="0" Grid.ColumnSpan="2"/>
-        <Border Grid.Row="3" Grid.Column="0" Grid.ColumnSpan="2" BorderThickness="2" BorderBrush="Black">
-            <Rectangle x:Name="previewRect">
-                <Rectangle.Fill>
-                    <SolidColorBrush/>
-                </Rectangle.Fill>
-            </Rectangle>
-        </Border>
-        <Label x:Name="previewLabel" Grid.Row="3" Grid.Column="2" VerticalAlignment="Bottom"/>
-        <Button x:Name="sendButton" Grid.Column="2" Grid.Row="4" Height="25" Width="75" Content="Send!" HorizontalAlignment="Right" VerticalAlignment="Bottom" Margin="0,0,20,10" Click="sendButton_Click"/>
-        <CheckBox x:Name="autosendBox" Grid.Column="2" VerticalAlignment="Center" Content="Autosending"/>
-        <Label x:Name="log" Grid.ColumnSpan="3" Grid.Row="5" Content="Log" VerticalAlignment="Top" Height="31"/>
-    </Grid>
-</Window>
